+++ packages/authentication/modules/user/useSimpleTokenUser/__tests__/useSimpleTokenUser.test.tsx
@@ -5,18 +5,38 @@ import {
   renderHook,
   waitFor,
 } from '@baseapp-frontend/test'
-import { axios } from '@baseapp-frontend/utils'
 
+import { UseQueryResult } from '@tanstack/react-query'
 import Cookies from 'js-cookie'
 
 import { IUser } from '../../../../types/user'
-import useSimpleTokenUser from '../index'
+import { IUseSimpleTokenUser } from '../types'
 import request from './fixtures/request.json'
 
-// @ts-ignore TODO: (BA-1081) investigate AxiosRequestHeaders error
-export const axiosMock = new MockAdapter(axios)
+interface IUseSimpleTokenUserResult<IUser> extends Omit<UseQueryResult<IUser, unknown>, 'data'> {
+  user?: IUser
+}
 
 describe('useSimpleTokenUser', () => {
+  let useSimpleTokenUser: <TUser extends Partial<IUser>>(
+    props?: IUseSimpleTokenUser<TUser>,
+  ) => IUseSimpleTokenUserResult<TUser>
+  let defaultTokenType: string | undefined
+  let axios: any
+  let axiosMock: MockAdapter
+
+  beforeAll(async () => {
+    defaultTokenType = process.env.NEXT_PUBLIC_TOKEN_TYPE
+    process.env.NEXT_PUBLIC_TOKEN_TYPE = 'simple'
+    axios = (await import('@baseapp-frontend/utils')).axios
+    axiosMock = new MockAdapter(axios)
+    useSimpleTokenUser = (await import('../index')).default as any
+  })
+
+  afterAll(() => {
+    process.env.NEXT_PUBLIC_TOKEN_TYPE = defaultTokenType
+  })
+
   test('should user be present for authenticated', async () => {
     ;(Cookies.get as CookiesGetByNameFn) = jest.fn(() => 'fake token')
 
@@ -80,4 +100,6 @@ describe('useSimpleTokenUser', () => {
     await waitFor(() => expect(Cookies.remove).toHaveBeenCalled())
     expect(result.current.user).not.toBeDefined()
   })
+
+  process.env.NEXT_PUBLIC_TOKEN_TYPE = defaultTokenType
 })