+++ packages/utils/functions/axios/createAxiosInstance/index.ts
@@ -1,4 +1,4 @@
-import _axios, { AxiosRequestConfig } from 'axios'
+import _axios from 'axios'
 import humps from 'humps'
 import Cookies from 'js-cookie'
 
@@ -6,11 +6,10 @@ import { SERVICES_WITHOUT_TOKEN } from '../../../constants/axios'
 import { ACCESS_COOKIE_NAME, REFRESH_COOKIE_NAME } from '../../../constants/cookie'
 import { LOGOUT_EVENT } from '../../../constants/events'
 import { TokenTypes } from '../../../constants/token'
-import { IJWTResponse } from '../../../types/jwt'
 import { eventEmitter } from '../../events'
 import { buildQueryString } from '../../string'
-
-const REFRESH_TOKEN_URL = '/auth/refresh'
+import { decodeJWT, isUserTokenValid } from '../../token'
+import { refreshAccessToken } from '../../token/refreshAccessToken'
 
 export const createAxiosInstance = ({
   returnData = true,
@@ -35,9 +34,22 @@ export const createAxiosInstance = ({
   instance.defaults.headers.put['Content-Type'] = contentType
 
   const requestInterceptorId = instance.interceptors.request.use(async (request) => {
-    const authToken = Cookies.get(cookieName)
+    let authToken = Cookies.get(cookieName)
 
     if (authToken) {
+      const isTokenValid =
+        tokenType === TokenTypes.jwt ? isUserTokenValid(decodeJWT(authToken)) : true
+      if (!isTokenValid) {
+        try {
+          authToken = await refreshAccessToken(cookieName, refreshCookieName)
+        } catch (error) {
+          if (eventEmitter.listenerCount(LOGOUT_EVENT)) {
+            eventEmitter.emit(LOGOUT_EVENT)
+          }
+          return Promise.reject(error)
+        }
+      }
+
       if (
         request.headers &&
         !request.headers.Authorization &&
@@ -67,22 +79,6 @@ export const createAxiosInstance = ({
       return returnData && response.data ? response.data : response
     },
     async (error) => {
-      if (
-        tokenType === TokenTypes.jwt &&
-        error.response?.status === 401 &&
-        error.config.url !== REFRESH_TOKEN_URL
-      ) {
-        try {
-          const originalRequest = error.config
-          // eslint-disable-next-line @typescript-eslint/no-use-before-define
-          return await refreshAccessToken(cookieName, refreshCookieName, originalRequest)
-        } catch (refreshError) {
-          if (eventEmitter.listenerCount(LOGOUT_EVENT)) {
-            eventEmitter.emit(LOGOUT_EVENT)
-          }
-        }
-      }
-
       if (error.response.data && error.response.headers?.['content-type'] === 'application/json') {
         const newError = { response: { data: {} } }
         newError.response.data = humps.camelizeKeys(error.response.data)
@@ -102,38 +98,3 @@ export const {
   requestInterceptorId: requestInterceptorIdForFiles,
   responseInterceptorId: responseInterceptorIdForFiles,
 } = createAxiosInstance({ file: true })
-
-// TODO: move this function to a separate file (we can't do it now because of a circular dependency)
-export const refreshAccessToken = async (
-  cookieName: string,
-  refreshCookieName: string,
-  originalRequest: AxiosRequestConfig,
-) => {
-  const refreshToken = Cookies.get(refreshCookieName)
-
-  if (!refreshToken) {
-    return Promise.reject(new Error('No refresh token'))
-  }
-
-  try {
-    const response = (await axios.post(REFRESH_TOKEN_URL, {
-      refresh: refreshToken,
-    })) as IJWTResponse
-
-    Cookies.set(cookieName, response.access, {
-      secure: process.env.NODE_ENV === 'production',
-    })
-
-    if (originalRequest.headers) {
-      // eslint-disable-next-line no-param-reassign
-      originalRequest.headers.Authorization = `Bearer ${response.access}`
-    }
-
-    return await axios(originalRequest)
-  } catch (error) {
-    Cookies.remove(cookieName)
-    Cookies.remove(refreshCookieName)
-
-    return Promise.reject(error)
-  }
-}