+++ packages/utils/functions/token/getToken/index.ts
@@ -1,10 +1,10 @@
 import ClientCookies from 'js-cookie'
-import { cookies as serverCookies } from 'next/headers'
 
 import { ACCESS_COOKIE_NAME } from '../../../constants/cookie'
 
-export const getToken = (cookieName = ACCESS_COOKIE_NAME) => {
+export const getToken = async (cookieName = ACCESS_COOKIE_NAME) => {
   if (typeof window === typeof undefined) {
+    const { cookies: serverCookies } = await import('next/headers')
     return serverCookies().get(cookieName)?.value
   }
   return ClientCookies.get(cookieName)