+++ packages/utils/functions/token/index.ts
@@ -1,3 +1,4 @@
 export * from './decodeJWT'
 export * from './getToken'
 export * from './isUserTokenValid'
+export * from './refreshAccessToken'