+++ packages/utils/hooks/useDebounce/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useRef } from 'react'
 
 import debounce from 'lodash/debounce'